fix paths and include test outputfile
