+++ tests/conftest.py
@@ -7,8 +7,10 @@ import pytest
 from astropy import units as u
 
 import simtools.io_handler
+import simtools.util.general as gen
 from simtools import db_handler
 from simtools.configuration.configurator import Configurator
+from simtools.corsika.corsika_output import CorsikaOutput
 from simtools.layout.layout_array import LayoutArray
 from simtools.model.telescope_model import TelescopeModel
 
@@ -246,3 +248,23 @@ def telescope_north_test_file():
 @pytest.fixture
 def telescope_south_test_file():
     return "data/layout/telescope_positions-South-TestLayout.ecsv"
+
+
+@pytest.fixture
+def corsika_output_file(io_handler):
+    corsika_output = gen.find_file(
+        "tests/resources/tel_output_10GeV-2-gamma-20deg-CTAO-South.dat",
+        io_handler.get_output_directory(dir_type="corsika_output", test=True),
+    )
+    return corsika_output
+
+
+@pytest.fixture
+def corsika_output_instance(db, io_handler, corsika_output_file):
+    # db.export_file_db(
+    #    db_name="test-data",
+    #    dest=io_handler.get_output_directory(dir_type="corsika_output", test=True),
+    #    file_name=test_file_name,
+    # )
+    # return CorsikaOutput(corsika_output_file)
+    return CorsikaOutput("tests/resources/tel_output_10GeV-2-gamma-20deg-CTAO-South.dat")
