+++ tests/unit_tests/corsika/test_corsika_output.py
@@ -2,6 +2,7 @@
 
 
 import copy
+from pathlib import Path
 
 import boost_histogram as bh
 import numpy as np
@@ -9,31 +10,9 @@ import pytest
 from astropy import units as u
 from astropy.io.misc import yaml
 
-import simtools.util.general as gen
 from simtools.corsika.corsika_output import CorsikaOutput, HistogramNotCreated
 
-test_file_name = "tel_output_10GeV-2-gamma-20deg-CTAO-South.dat"
-# test_file_name = "tel_output.dat"
-
-
-@pytest.fixture
-def corsika_output_file(io_handler):
-    corsika_output = gen.find_file(
-        test_file_name,
-        io_handler.get_output_directory(dir_type="corsika_output", test=True),
-    )
-    return corsika_output
-
-
-@pytest.fixture
-def corsika_output_instance(db, io_handler, corsika_output_file):
-    # db.export_file_db(
-    #    db_name="test-data",
-    #    dest=io_handler.get_output_directory(dir_type="corsika_output", test=True),
-    #    file_name=test_file_name,
-    # )
-    # return CorsikaOutput(corsika_output_file)
-    return CorsikaOutput(test_file_name)
+test_file_name = "tests/resources/tel_output_10GeV-2-gamma-20deg-CTAO-South.dat"
 
 
 @pytest.fixture
@@ -48,7 +27,7 @@ def test_file_exists(corsika_output_file):
 
 
 def test_init(corsika_output_instance):
-    assert corsika_output_instance.input_file.name == test_file_name
+    assert corsika_output_instance.input_file.name == Path(test_file_name).name
     with pytest.raises(FileNotFoundError):
         CorsikaOutput("wrong_file_name")
     assert len(corsika_output_instance.event_information) > 15
